Create README.md
                      Totot Traditional Hotel
Totot Traditional Hotel is the most well-known and excellent traditional eatery in Addis Abeba.
and since we all need to know how the traditional foods and drinks there are prepared 
my goal is to build a website that can display the traditional foods and drinks there.
posted either weekly or daily by the master chief, who serves as the website administrator or admin, 
and that includes all the information about how the foods and drinks are prepared, their ingredients, and 
taking the time to provide all the necessary details. Using Node.js and Express, I created the backend.